[] Use flowbit more and more... oricu e de cacat
