+++ package.json
@@ -27,5 +27,11 @@
 		"typescript": "5.0.4",
 		"vite": "4.3.0"
 	},
-	"type": "module"
+	"type": "module",
+	"dependencies": {
+		"@popperjs/core": "^2.11.8",
+		"classnames": "^2.3.2",
+		"flowbite": "^1.6.5",
+		"flowbite-svelte": "^0.37.2"
+	}
 }
