+++ src/lib/header/index.svelte
@@ -5,9 +5,3 @@
 <div class="flex items-center justify-center w-full layout">
     <TopNavigation/>
 </div>
-
-<style>
-    .layout {
-        background:red;
-    }
-</style>
