+++ src/lib/top-navigation/index.svelte
@@ -1,21 +1,28 @@
 <script>
     import NerdsLogo from './nerds_logo.svg?component'
-    import Menu from './menu.svelte'
+    import {Navbar, NavBrand, NavHamburger, NavLi, NavUl} from "flowbite-svelte";
+    import LanguageSelector from "$lib/top-navigation/LanguageSelector.svelte";
+    const menuItems = [
+        {link:'case-studies', text:"Case Studies"},
+        {link:'services', text:"Services"},
+        {link:'products', text:"Products"},
+        {link:'carriers', text:"Carriers"},
+        {link:'blog', text:"Blog"},
+        {link:'contact', text:"Contact"}
+    ]
 </script>
 
-<nav class="bg-white w-full border-gray-200 dark:bg-gray-900">
-    <div class="bg-white border-gray-200 dark:bg-gray-900">
-        <div class="max-w-screen-xl flex flex-wrap items-center justify-between mx-auto p-4">
-            <a href="/#" class="flex items-center">
-                <NerdsLogo/>
-            </a>
-            <button data-collapse-toggle="navbar-default" type="button" class="inline-flex items-center p-2 ml-3 text-sm text-gray-500 rounded-lg md:hidden hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-gray-200 dark:text-gray-400 dark:hover:bg-gray-700 dark:focus:ring-gray-600" aria-controls="navbar-default" aria-expanded="false">
-                <span class="sr-only">Open main menu</span>
-                <svg class="w-6 h-6" aria-hidden="true" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path fill-rule="evenodd" d="M3 5a1 1 0 011-1h12a1 1 0 110 2H4a1 1 0 01-1-1zM3 10a1 1 0 011-1h12a1 1 0 110 2H4a1 1 0 01-1-1zM3 15a1 1 0 011-1h12a1 1 0 110 2H4a1 1 0 01-1-1z" clip-rule="evenodd"></path></svg>
-            </button>
-            <div class="hidden w-full md:block md:w-auto" id="navbar-default">
-                <Menu/>
-            </div>
-        </div>
-    </div>
-</nav>
+<Navbar let:hidden let:toggle>
+    <NavBrand href="/">
+       <NerdsLogo/>
+    </NavBrand>
+    <NavHamburger on:click={toggle} />
+    <NavUl {hidden}>
+        {#each menuItems as menuItem}
+            <NavLi href="{menuItem.to}">{menuItem.text}</NavLi>
+        {/each}
+    </NavUl>
+    <LanguageSelector/>
+</Navbar>
+
+
